+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.1.2]
+
+* fix isOnTapEnabled
+* add ability to add box shadow to element
+* add on drag start action
+
 ## [0.1.1]
 
 * fix reverse animation